+++ frontend/src/components/ads/AdDelete.tsx
@@ -3,7 +3,8 @@ import { useMutation } from "@apollo/client";
 import CardActions from "@mui/material/CardActions";
 import Button from "@mui/material/Button";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { mutationDeleteAd, queryAllAds } from "../../graphql/Ads";
+import { mutationDeleteAd } from "@/graphql/ads/mutationDeleteAd";
+import { queryAllAds } from "@/graphql/ads/queryAllAds";
 
 type DeleteAdProps = {
   id: number;
