+++ frontend/src/components/search/Search.tsx
@@ -2,10 +2,10 @@ import React, { useState } from "react";
 import { AdTags, AdsTypes } from "@/types/AdTypes";
 import { useLazyQuery } from "@apollo/client";
 import { TextField, Box, Button, Grid, useTheme } from "@mui/material";
-import { FilterAlt, FilterAltOff, Tag } from "@mui/icons-material";
+import { FilterAlt, FilterAltOff } from "@mui/icons-material";
 import { Toaster } from "react-hot-toast";
 import { PATH_IMAGE } from "@/api/configApi";
-import { queryAllAds } from "../../graphql/Ads";
+import { queryAllAds } from "../../graphql/ads/queryAllAds";
 import GpsAndRadius from "./components/GpsAndRadius";
 import { VariablesColors } from "@/styles/Variables.colors";
 import { showToast } from "../utils/toastHelper";
