+++ frontend/src/graphql/ads/mutationCreateAd.ts
@@ -1,0 +1,9 @@
+import { gql } from "@apollo/client";
+
+export const mutationCreateAd = gql`
+  mutation adCreate($data: AdCreateInput!) {
+    item: adCreate(data: $data) {
+      id
+    }
+  }
+`;
