+++ frontend/src/graphql/ads/mutationDeleteAd.ts
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export const mutationDeleteAd = gql`
+  mutation adDelete($adDeleteId: ID!) {
+    item: adDelete(id: $adDeleteId) {
+      id
+      title
+    }
+  }
+`;
