+++ frontend/src/pages/ads/[id]/edit.tsx
@@ -2,7 +2,7 @@ import LayoutFull from "@/components/layout/LayoutFull";
 import { AdTypes } from "@/types/AdTypes";
 import AdForm from "@/components/ads/adForm/AdForm";
 import AdCard from "@/components/ads/AdCard";
-import { queryAdById } from "@/graphql/Ads";
+import { queryAdById } from "@/graphql/ads/queryAdById";
 import { useQuery } from "@apollo/client";
 import { useRouter } from "next/router";
 import { Box, Typography } from "@mui/material";
