+++ frontend/src/pages/ads/[id]/index.tsx
@@ -1,7 +1,7 @@
 import LayoutFull from "@/components/layout/LayoutFull";
 import { useRouter } from "next/router";
 import { AdTypes } from "@/types/AdTypes";
-import { queryAdById } from "@/graphql/Ads";
+import { queryAdById } from "@/graphql/ads/queryAdById";
 import { useQuery } from "@apollo/client";
 import IconBreadcrumbs from "@/components/breadcrumbs/Breadcrumbs";
 import LoadingApp from "@/styles/LoadingApp";
